requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
PyMySQL==1.1.0
APScheduler==3.10.4
requests==2.31.0
reportlab==4.0.7
python-dotenv==1.0.0

